projects/01_ecommerce_analytics_platform/requirements.txt
# ============================================================
# SNOWBRIX E-COMMERCE PLATFORM — Python Dependencies
# ============================================================

# --- Snowflake ---
snowflake-connector-python==3.6.0
snowflake-sqlalchemy==1.5.1

# --- Source Database ---
psycopg2-binary==2.9.9

# --- Stripe API ---
stripe==8.0.0

# --- HTTP / Retry ---
requests==2.31.0
tenacity==8.2.3

# --- Data Processing ---
pandas==2.2.0
pyarrow==15.0.0

# --- dbt ---
dbt-core==1.7.8
dbt-snowflake==1.7.3

# --- Dashboard ---
streamlit==1.30.0
plotly==5.18.0
altair==5.2.0

# --- Orchestration (if running Airflow locally without Docker) ---
# apache-airflow==2.8.1
# apache-airflow-providers-snowflake==5.3.0
# apache-airflow-providers-slack==8.5.0

# --- Utilities ---
python-dotenv==1.0.0
pyyaml==6.0.1
click==8.1.7
